requirements.txt
Pillow
colorama
av
numpy
opencv-python
cython
setuptools